refactor: Abstracted two methods into StepRunner
