+++ src/popper/runner.py
@@ -230,6 +230,74 @@ class StepRunner(object):
             )
         return step_env
 
+    def _get_build_info(self, step):
+        """Parses the `uses` attribute and returns build information needed.
+
+        Args:
+            step(dict): dict with step data
+
+        Returns:
+            (str, str, str, str): bool (build), image, tag, Dockerfile
+        """
+        build = True
+        img = None
+        build_ctx_path = None
+        img_full = None
+
+        if (
+            "docker://" in step.uses
+            or "shub://" in step.uses
+            or "library://" in step.uses
+        ):
+            img_full = step.uses
+            img = step.uses.replace("docker://", "")
+            if ":" in img:
+                (img, tag) = img.split(":")
+            else:
+                tag = "latest"
+            build = False
+        elif "./" in step.uses:
+            img = pu.sanitized_name(step.id, "step").lower()
+            tag = self._config.git_sha_short if self._config.git_sha_short else "na"
+            build_ctx_path = os.path.join(self._config.workspace_dir, step.uses)
+        else:
+            _, service, user, repo, step_dir, version = scm.parse(step.uses)
+            wf_cache_dir = os.path.join(self._config.cache_dir, self._config.wid)
+            repo_dir = os.path.join(wf_cache_dir, service, user, repo)
+            img = f"{user}/{repo}".lower()
+            tag = version
+            build_ctx_path = os.path.join(repo_dir, step_dir)
+
+        return (build, img_full, img, tag, build_ctx_path)
+
+    def _update_with_engine_config(self, container_args):
+        """Given container arguments, it extends it so it includes options
+        obtained from the popper.config.Config.engine_opts property.
+        """
+        update_with = self._config.engine_opts
+        if not update_with:
+            return
+
+        if container_args.get("volumes"):
+            container_args["volumes"] = [
+                *container_args["volumes"],
+                *update_with.get("volumes", list()),
+            ]
+
+        if container_args.get("bind"):
+            container_args["bind"] = [
+                *container_args["bind"],
+                *update_with.get("bind", list()),
+            ]
+
+        if update_with.get("environment"):
+            for k, v in update_with.get("environment", dict()).items():
+                container_args["environment"].update({k: v})
+
+        for k, v in update_with.items():
+            if k not in container_args.keys():
+                container_args[k] = update_with[k]
+
     def stop_running_tasks(self):
         raise NotImplementedError("Needs implementation in derived classes.")
