+++ src/test/test_runner_host.py
@@ -137,6 +137,7 @@ class TestHostDockerRunner(PopperTest):
             self.assertEqual(c.status, "created")
             c.remove()
 
+    @unittest.skipIf(os.environ.get("ENGINE", "docker") != "docker", "ENGINE != docker")
     def test_find_container(self):
         config = ConfigLoader.load()
         step = Box(
@@ -260,56 +261,6 @@ class TestHostDockerRunner(PopperTest):
                     "ports": {"8888/tcp": 8888},
                 },
             )
-
-    @unittest.skipIf(os.environ.get("ENGINE", "docker") != "docker", "ENGINE != docker")
-    def test_get_build_info(self):
-        step = Box(
-            {"uses": "popperized/bin/sh@master", "args": ["ls"], "id": "one",},
-            default_box=True,
-        )
-        with DockerRunner(init_docker_client=False) as dr:
-            build, img, tag, build_ctx_path = dr._get_build_info(step)
-            self.assertEqual(build, True)
-            self.assertEqual(img, "popperized/bin")
-            self.assertEqual(tag, "master")
-            self.assertTrue(f"{os.environ['HOME']}/.cache/popper" in build_ctx_path)
-            self.assertTrue("github.com/popperized/bin/sh" in build_ctx_path)
-
-            step = Box(
-                {
-                    "uses": "docker://alpine:3.9",
-                    "runs": ["sh", "-c", "echo $FOO > hello.txt ; pwd"],
-                    "env": {"FOO": "bar"},
-                    "id": "1",
-                },
-                default_box=True,
-            )
-
-        with DockerRunner(init_docker_client=False) as dr:
-            build, img, tag, build_ctx_path = dr._get_build_info(step)
-            self.assertEqual(build, False)
-            self.assertEqual(img, "alpine")
-            self.assertEqual(tag, "3.9")
-            self.assertEqual(build_ctx_path, None)
-
-        step = Box({"uses": "./", "args": ["ls"], "id": "one",}, default_box=True,)
-        conf = ConfigLoader.load(workspace_dir="/tmp")
-        with DockerRunner(init_docker_client=False, config=conf) as dr:
-            build, img, tag, build_ctx_path = dr._get_build_info(step)
-            self.assertEqual(build, True)
-            self.assertEqual(img, "popper_one_step")
-            self.assertEqual(tag, "na")
-            self.assertEqual(build_ctx_path, f"{os.path.realpath('/tmp')}/./")
-
-        # test within a git repo
-        repo = self.mk_repo()
-        conf = ConfigLoader.load(workspace_dir=repo.working_dir)
-        with DockerRunner(init_docker_client=False, config=conf) as dr:
-            build, img, tag, build_ctx_path = dr._get_build_info(step)
-            self.assertEqual(build, True)
-            self.assertEqual(img, "popper_one_step")
-            self.assertEqual(tag, scm.get_sha(repo, short=7))
-            self.assertEqual(build_ctx_path, f"{os.path.realpath(repo.working_dir)}/./")
 
     @unittest.skipIf(os.environ.get("ENGINE", "docker") != "docker", "ENGINE != docker")
     def test_docker_basic_run(self):
@@ -461,56 +412,6 @@ class TestHostPodmanRunner(PopperTest):
             self.assertEqual(c_status, "configured\n")
             cmd = ["podman", "container", "rm", c]
             HostRunner._exec_cmd(cmd, logging=False)
-
-    @unittest.skipIf(os.environ.get("ENGINE", "docker") != "podman", "ENGINE != podman")
-    def test_get_build_info(self):
-        step = Box(
-            {"uses": "popperized/bin/sh@master", "args": ["ls"], "id": "one",},
-            default_box=True,
-        )
-        with PodmanRunner(init_podman_client=False) as pr:
-            build, img, tag, build_sources = pr._get_build_info(step)
-            self.assertEqual(build, True)
-            self.assertEqual(img, "popperized/bin")
-            self.assertEqual(tag, "master")
-            self.assertTrue(f"{os.environ['HOME']}/.cache/popper" in build_sources)
-            self.assertTrue("github.com/popperized/bin/sh" in build_sources)
-
-            step = Box(
-                {
-                    "uses": "docker://alpine:3.9",
-                    "runs": ["sh", "-c", "echo $FOO > hello.txt ; pwd"],
-                    "env": {"FOO": "bar"},
-                    "id": "1",
-                },
-                default_box=True,
-            )
-
-        with PodmanRunner(init_podman_client=False) as pr:
-            build, img, tag, build_sources = pr._get_build_info(step)
-            self.assertEqual(build, False)
-            self.assertEqual(img, "alpine")
-            self.assertEqual(tag, "3.9")
-            self.assertEqual(build_sources, None)
-
-        # step = Box({"uses": "./", "args": ["ls"], "id": "one",}, default_box=True,)
-        # conf = ConfigLoader.load(workspace_dir="/tmp")
-        # with PodmanRunner(init_podman_client=False, config=conf) as pr:
-        #     build, img, tag, build_ctx_path = pr._get_build_info(step)
-        #     self.assertEqual(build, True)
-        #     self.assertEqual(img, "popper_one_step")
-        #     self.assertEqual(tag, "na")
-        #     self.assertEqual(build_ctx_path, f"{os.path.realpath('/tmp')}/./")
-
-        # # test within a git repo
-        # repo = self.mk_repo()
-        # conf = ConfigLoader.load(workspace_dir=repo.working_dir)
-        # with PodmanRunner(init_podman_client=False, config=conf) as pr:
-        #     build, img, tag, build_ctx_path = pr._get_build_info(step)
-        #     self.assertEqual(build, True)
-        #     self.assertEqual(img, "popper_one_step")
-        #     self.assertEqual(tag, scm.get_sha(repo, short=7))
-        #     self.assertEqual(build_ctx_path, f"{os.path.realpath(repo.working_dir)}/./")
 
     @unittest.skipIf(os.environ.get("ENGINE", "docker") != "podman", "ENGINE != podman")
     def test_get_container_kwargs(self):
@@ -768,37 +669,6 @@ exec /bin/bash "$@"''',
                     "--ipc",
                 ],
             )
-
-    @unittest.skipIf(
-        os.environ.get("ENGINE", "docker") != "singularity", "ENGINE != singularity"
-    )
-    def test_get_build_info(self):
-        step = Box(
-            {"uses": "popperized/bin/sh@master", "args": ["ls"], "name": "one",},
-            default_box=True,
-        )
-        with SingularityRunner() as sr:
-            build, img, build_ctx_path = sr._get_build_info(step)
-            self.assertEqual(build, True)
-            self.assertEqual(img, "popperized/bin")
-            self.assertTrue(f"{os.environ['HOME']}/.cache/popper" in build_ctx_path)
-            self.assertTrue("github.com/popperized/bin/sh" in build_ctx_path)
-
-            step = Box(
-                {
-                    "uses": "docker://alpine:3.9",
-                    "runs": ["sh", "-c", "echo $FOO > hello.txt ; pwd"],
-                    "env": {"FOO": "bar"},
-                    "name": "1",
-                },
-                default_box=True,
-            )
-
-        with SingularityRunner() as sr:
-            build, img, build_ctx_path = sr._get_build_info(step)
-            self.assertEqual(build, False)
-            self.assertEqual(img, "docker://alpine:3.9")
-            self.assertEqual(build_ctx_path, None)
 
     @unittest.skipIf(
         os.environ.get("ENGINE", "docker") != "singularity", "ENGINE != singularity"
